add struct handling plumbing, support for non-owning types
This adds new builtin types STRUCT and ENUM, which will be used
for types to refer to defined struct/enum declarations. It also
adds more code for typedefs and adds non-owning flag for types
so that they can refer to a pointer that's not theirs. This is
newly used for global function references, to reduce allocations.

Also add infix syntax for typedefs as allowed by C.