unify all cdata's alignment and implement scalar cdata
This ensures that *all* cdata, including functions, pointers,
scalars and later large values (e.g. structs) will have
consistent alignment, which makes it all safe to cast
between.

It also adds scalar cdata, which are basically immutable boxed
values and generally not that useful, but they come in handy
once in a while.