fix redefinition errors for anon-structs-inside-anon-structs
The problem here is that when you had a nested anonymous struct
declaration like this:

```
typedef struct {
    struct { ... } bar;
} foo;
```

By the time the name for the inner struct was requested, the outer
struct was still in the middle of parsing, i.e. it was not written
into the storage; and since `request_name` was based on the number
of records already written in storage, the outer structure would
get the same name as the inner structure.

To deal with this, do not use the number of elements written, but
instead keep a separate counter inside the declaration store,
incremented each time a name is requested.

Fixes https://github.com/q66/cffi-lua/issues/11