fix behavior of unary operator metatype methods on newer lua
Newer versions of lua pass two arguments to metamethods instead of
just one to simplify the implementation; that means attempting to
implement the logic by inserting the function to the top of the
stack and then calling with 1 argument would not work and instead
would try to call the object themselves, as it would be present
just below the top.

Instead, push the object above the function (which was previously
at the top) and call regardless of previously present args.

Fixes https://github.com/q66/cffi-lua/issues/31